Nov 27 Wed

+++ src/main/java/com/rs/employer/service/account/JournalEntryDetailService.java
@@ -3,12 +3,18 @@ package com.rs.employer.service.account;
 import com.rs.employer.dao.account.ChartOfAccountRepository;
 import com.rs.employer.dao.account.JournalEntryDetailRepository;
 import com.rs.employer.dao.account.JournalEntryRepository;
+import com.rs.employer.dao.customer.CustomerRepo;
 import com.rs.employer.dto.Request.JournalEntryDetailRequestDTO;
+import com.rs.employer.globalexception.AppException;
+import com.rs.employer.globalexception.ErrorCode;
 import com.rs.employer.mapper.JournalEntryDetailMapper;
 import com.rs.employer.model.account.JournalEntryDetail;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,17 +24,19 @@ public class JournalEntryDetailService {
     private final JournalEntryDetailMapper journalEntryDetailMapper;
     private final ChartOfAccountRepository chartOfAccountRepository;
     private final JournalEntryRepository journalEntryRepository;
+    private final CustomerRepo customerRepo;
 
     @Autowired
     public JournalEntryDetailService(JournalEntryDetailRepository journalEntryDetailRepository,
                                      JournalEntryDetailMapper journalEntryDetailMapper,
                                      ChartOfAccountRepository chartOfAccountRepository,
-                                     JournalEntryRepository journalEntryRepository
-    ) {
+                                     JournalEntryRepository journalEntryRepository,
+                                     CustomerRepo customerRepo) {
         this.journalEntryDetailRepository = journalEntryDetailRepository;
         this.journalEntryDetailMapper = journalEntryDetailMapper;
         this.chartOfAccountRepository = chartOfAccountRepository;
         this.journalEntryRepository = journalEntryRepository;
+        this.customerRepo = customerRepo;
     }
 
 
@@ -43,7 +51,13 @@ public class JournalEntryDetailService {
 
 
     public JournalEntryDetail createJournalEntryDetail(JournalEntryDetailRequestDTO journalEntryRequest) {
+        String authenticator  = SecurityContextHolder.getContext().getAuthentication().getName();
+        System.out.println("Người dùng là "  +authenticator);
         JournalEntryDetail journalEntryDetail = journalEntryDetailMapper.map(journalEntryRequest);
+        if(!customerRepo.existsByUsername(authenticator))
+            throw new AppException(ErrorCode.USER_NOTFOUND);
+        journalEntryDetail.setCreateDate(LocalDate.now());
+        journalEntryDetail.setCreateBy(authenticator);
         journalEntryDetail.setJournalEntry(journalEntryRepository.findById(journalEntryRequest.getJournalEntryId()).get());
         journalEntryDetail.setAccount(chartOfAccountRepository.findById(journalEntryRequest.getChartOfAccountCode()).get());
         return journalEntryDetailRepository.save(journalEntryDetail);
@@ -51,8 +65,9 @@ public class JournalEntryDetailService {
 
 
     public JournalEntryDetail updateJournalEntryDetail(Integer id, JournalEntryDetailRequestDTO journalEntryRequest) {
-        if (!journalEntryDetailRepository.existsById(id)) {
-            return null;
+        String authentication = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!journalEntryDetailRepository.existsById(id) || customerRepo.existsByUsername(authentication)) {
+            throw new AppException(ErrorCode.UNCATEGORIZE_EXCEPTION);
         }
         JournalEntryDetail journalEntryDetail = journalEntryDetailMapper.map(journalEntryRequest);
         journalEntryDetail.setEntryDetailId(id);
@@ -61,6 +76,8 @@ public class JournalEntryDetailService {
 
 
     public void deleteJournalEntryDetail(Integer id) {
+        if(journalEntryRepository.existsById(id))
         journalEntryDetailRepository.deleteById(id);
+        else throw new AppException(ErrorCode.UNCATEGORIZE_EXCEPTION);
     }
 }
